lift update to route
